Move CAS1 migration jobs into cas1 package
